+++ src/main/java/me/xginko/villageroptimizer/modules/gameplay/PreventOptimizedTargeting.java
@@ -1,9 +1,9 @@
 package me.xginko.villageroptimizer.modules.gameplay;
 
-import com.cryptomorin.xseries.XEntityType;
 import me.xginko.villageroptimizer.modules.VillagerOptimizerModule;
 import me.xginko.villageroptimizer.wrapper.WrappedVillager;
 import org.bukkit.entity.Entity;
+import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Mob;
 import org.bukkit.entity.Villager;
 import org.bukkit.event.EventHandler;
@@ -39,7 +39,7 @@ public class PreventOptimizedTargeting extends VillagerOptimizerModule implement
         final Entity target = event.getTarget();
         if (
                 target != null
-                && target.getType() == XEntityType.VILLAGER.get()
+                && target.getType() == EntityType.VILLAGER
                 && wrapperCache.get((Villager) target, WrappedVillager::new).isOptimized()
         ) {
             event.setTarget(null);
@@ -52,7 +52,7 @@ public class PreventOptimizedTargeting extends VillagerOptimizerModule implement
         final Entity target = event.getTargetEntity();
         if (
                 target != null
-                && target.getType() == XEntityType.VILLAGER.get()
+                && target.getType() == EntityType.VILLAGER
                 && wrapperCache.get((Villager) target, WrappedVillager::new).isOptimized()
         ) {
             event.setCancelled(true);
@@ -62,7 +62,7 @@ public class PreventOptimizedTargeting extends VillagerOptimizerModule implement
     @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     private void onEntityAttackVillager(EntityDamageByEntityEvent event) {
         if (
-                event.getEntityType() == XEntityType.VILLAGER.get()
+                event.getEntityType() == EntityType.VILLAGER
                 && event.getDamager() instanceof Mob
                 && wrapperCache.get((Villager) event.getEntity(), WrappedVillager::new).isOptimized()
         ) {